Fix bugs 1

+++ data/src/main/java/com/example/data/db/tables/movie_tables/SavedMovieEntity.kt
@@ -10,7 +10,7 @@ class SavedMovieEntity(
     val releaseData: String?,
     val popularityScore: String?,
     val movieName: String?,
-    val rating: Double?,
+    val rating: Float?,
     val poster: String?,
     val overview: String?,
     val backdropPoster: String?,
