+++ domain/src/main/java/com/example/domain/models/MovieWithDetailsModel.kt
@@ -7,7 +7,7 @@ data class MovieWithDetailsModel(
     val releaseData: String?,
     val popularityScore: String?,
     val movieName: String?,
-    val rating: Double?,
+    val rating: Float?,
     val poster: String?,
     val overview: String?,
     val backdropPoster: String?,
